Problem 96: identify a symmetric binary tree
Let us define a binary tree as "symmetric" if the left half of the tree
is the mirror image of the right half of the tree.  Write a predicate to
determine whether or not a given binary tree is symmetric.
See problem 95 for a reminder of the tree representation we're using.